requirements.txt
flask
gunicorn
flask_sqlalchemy
marshmallow_sqlalchemy
flask_marshmallow
psycopg2==2.8
pyjwt